requirements.txt
google-genai
flask
flask-sock
python-dotenv
gunicorn
pydub
ffmpeg-python